+++ packages/api/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.77.3](https://github.com/betagouv/api-subventions-asso/compare/v0.77.2...v0.77.3) (2026-01-13)
+
+### Bug Fixes
+
+- **api, front:** [#3773](https://github.com/betagouv/api-subventions-asso/issues/3773) truncate scdl error in deposit log ([b022bf8](https://github.com/betagouv/api-subventions-asso/commit/b022bf8d997ec7441b8eb098f590188a4f1ad281))
+- **api, front:** [#3773](https://github.com/betagouv/api-subventions-asso/issues/3773) truncate scdl error in deposit log ([0b8b496](https://github.com/betagouv/api-subventions-asso/commit/0b8b49682ec50dda923cd7f6ce9f044c1973fe69))
+
 ## [0.77.1](https://github.com/betagouv/api-subventions-asso/compare/v0.77.0...v0.77.1) (2026-01-05)
 
 ### Bug Fixes